Improve logging in case of call failure
Previously, we were logging the retry attempt before we even evaluated
whether or not we are going to retry again. Instead, we now compute
the number of failed attempts and either log and bail from the function
or inform the user that we are going to retry the call.

We also include the name of the failed call and the reason why it failed
in case we are retrying it. Once we have exhausted the the number of
retries, we return all errors anyway so there is no reason to log them.
It is up to the caller to perform error handling.