Enable TCP support without socks5 or SSL
Currently users of the library must configure SSL or a socks5 proxy even
if they only want a plain TCP stream. This is contrary to the comment in
`lib.rs` that suggests `features=minimal` may be used to get TCP
only (will not currently build with `--no-default-features`).

There is no need to force plaintext TCP stream users to enable SSL
and/or proxy features.

Enable `Client` to be used for plaintext TCP stream without requiring
any of the features `use-openssl`, `use-rustls`, or `proxy`.