Test to showcase high memory consumption of `batch_script_get_history`
When running `batch_script_get_history` on a Linux machine with processing many scripts the memory consumption is enormous and does not go down.
This test showcases this behavior. When triggering the test with 7000 scripts we see a memory consumption of about 4GB.
This problem could be reproduced on multiple Linux machine, but not on MacOS.

Note that you can add a loop to re-trigger the batching to see that the memory does not get freed.