Add `Batch::raw` and improve docs
Being able to add raw requests to `Batch` makes sense from an API
standpoint because we already allow raw non-batched requests. This is
also useful when the electrum server API gets an updated version and our
client is unable to keep up.

Additional to this, I have improved the documentation and made `Call`
private (since `Call` is never used externally).