build: Clean up "proxy" feature
Before, the `client` module was gated on the "proxy" feature,
which caused unnecessary coupling since proxy is only
required for the Socks5 client type.

This change removes the client module's dependency on the "proxy"
feature in favor of more precise feature gating of the Socks5 client
type and related functions calling into the `socks` module.

Note that `Client` still requires a TLS backend to be enabled
by one of "use-openssl", "use-rustls" or "use-rustls-ring".